Doxygen

+++ src/main/MischbaresRezeptbuch.cpp
@@ -71,7 +71,7 @@ void MischbaresRezeptbuch::setZutatenVerwalter(VorhandeneZutaten * zv) {
 }
 
 void MischbaresRezeptbuch::allRecipesMixableTrue(){
-  for(size_t i = 0; i < getNumberOfRecipes(); i++){
+  for(int i = 0; i < getNumberOfRecipes(); i++){
     getRecipe(i)->isMixable = true;
   }
